+++ packages/client/src/pages/Register/types.ts
@@ -1,0 +1,8 @@
+export interface RegisterForm {
+  name: string
+  lastName: string
+  login: string
+  email: string
+  password: string
+  phone: string
+}
